refactor: centralize Vditor init
